It compiles!

+++ ProyectosSDL/HolaSDL/ArkanoidObject.cpp
@@ -3,5 +3,5 @@
 #include"ArkanoidObject.h"
 
 void ArkanoidObject::render() const {
-	_texture->render(SDL_Rect{ _position.x,_position.y,_size.x,_size.y });
+	_texture->render(SDL_Rect{ (int)_position.x, (int)_position.y, (int)_size.x, (int)_size.y });
 }
